requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.6.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.3
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
cytoolz==0.12.2
decorator==5.1.1
Django==4.2
django-cors-headers==3.14.0
django-extensions==3.2.1
django-rest-registration==0.8.2
djangorestframework==3.14.0
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
exceptiongroup==1.1.3
executing==2.0.1
frozenlist==1.4.0
hexbytes==0.3.1
idna==3.4
ipython==8.17.2
jedi==0.19.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
matplotlib-inline==0.1.6
multidict==6.0.4
parsimonious==0.9.0
parso==0.8.3
Pillow==9.5.0
prompt-toolkit==3.0.41
protobuf==4.25.0
pure-eval==0.2.2
py-solc==3.2.0
pycryptodome==3.19.0
Pygments==2.16.1
pytz==2023.3
pyunormalize==15.1.0
pywin32==306
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.12.0
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.3
stack-data==0.6.3
toolz==0.12.0
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.10
web3==6.11.3
websockets==12.0
yarl==1.9.2
